requirements.txt
setuptools>=69.0.2
wheel>=0.42.0
packaging>=23.2
Flask==3.0.0
flask-restx>=1.3.0
Pillow==10.2.0
numpy>=1.26.3
transformers>=4.36.0
python-dotenv==1.0.0
opencv-python-headless>=4.8.0
torch>=2.1.0
tqdm>=4.66.1 